Moving legacy shit where it can be looked at without causing circular import
