+++ router/admin/ssh_secret.go
@@ -10,7 +10,7 @@ import (
 	"github.com/level11consulting/ocelot/storage"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"github.com/level11consulting/ocelot/secret"
+	"github.com/level11consulting/ocelot/secret/legacy"
 	"github.com/level11consulting/ocelot/router/admin/anycred"
 	"github.com/level11consulting/ocelot/server/config"
 )
@@ -54,7 +54,7 @@ func (g *SshSecretAPI) SetSSHCreds(ctx context.Context, creds *pb.SSHKeyWrapper)
 	}
 	// no validation necessary, its a file upload
 
-	err := secret.SetupRCCCredentials(g.RemoteConfig, g.Storage, creds)
+	err := legacy.SetupRCCCredentials(g.RemoteConfig, g.Storage, creds)
 	if err != nil {
 		if _, ok := err.(*pb.ValidationErr); ok {
 			return &empty.Empty{}, status.Error(codes.InvalidArgument, "SSH Creds Upload failed validation. Errors are: "+err.Error())
